requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Vector stores
chromadb>=0.4.0
faiss-cpu>=1.7.0

# Embeddings and ML
sentence-transformers>=2.2.0
transformers>=4.20.0
torch>=1.12.0

# API clients
google-genai>=0.1.0
nvidia-nim>=0.1.0
groq>=0.4.0

# Database
sqlalchemy>=1.4.0
pymysql>=1.0.0
mysql-connector-python>=8.0.0

# Text processing
langchain>=0.1.0
tiktoken>=0.4.0
pymupdf>=1.20.0
python-docx>=0.8.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Utilities
python-dotenv>=0.19.0
tqdm>=4.62.0
pyyaml>=6.0
requests>=2.25.0

# Development
pytest>=6.0.0
jupyter>=1.0.0
ipykernel>=6.0.0
black>=22.0.0
flake8>=4.0.0

# Logging
loguru>=0.6.0
